management/requirements.txt
Django<4
psycopg2>=2.9
djangorestframework>=3.13
djangorestframework_simplejwt
djoser
django-environ
django-cors-headers
django-filter
pytest==7.0.1
isort>=5.10.1
black>=22.1.0
elasticsearch-dsl>=6.0, <7.0
django-elasticsearch-dsl>=6.0,<7.0
django-polymorphic
pytest-django
requests==2.27
requests-mock==1.9.3
pydantic[dotenv]==1.9
starlette==0.20.0
requests-toolbelt==0.9.1
requests_cache
django-rest-polymorphic==0.1.9
jsonfield
pyyaml==6.0
uritemplate
mock
parameterized==0.8.1
pytest-django
pytest
pandas==1.4.2
numpy
